Decompile almost all of cable link code

+++ include/data/cable_link_data.h
@@ -4,6 +4,6 @@
 #include "types.h"
 
 extern const u8 sTransferData_8754bd0[];
-extern const u8 sTransferData_87566c4[];
+extern const u32 sTransferData_87566c4[];
 
 #endif /* CABLE_LINK_DATA_H */
